requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro-datasets>=3.0
kedro-datasets[pandas]>=7.0.0,<8.0.0
kedro-viz>=6.7.0
kedro[jupyter]~=0.19.12
notebook
scikit-learn~=1.5.1
matplotlib>=3.10.3
seaborn>=0.13.2
kagglehub>=0.3.12
unidecode>=1.3.6
geopandas>=0.14.0
datashader>=0.15.2
setuptools; python_version >= "3.10.16"
nbstripout>=0.8.1
haversine>=2.9.0
torch>=2.2.2
pysentimiento>=0.3.0
lightgbm>=4.6.0